+++ packages/pluggableWidgets/2G-datagrid-web/src/Datagrid.editorConfig.ts
@@ -9,6 +9,7 @@ import {
     Properties,
     RowLayoutProps,
     StructurePreviewProps,
+    TextProps,
     transformGroupsIntoTabs
 } from "@mendix/piw-utils-internal";
 import { ColumnsPreviewType, DatagridPreviewProps } from "../typings/DatagridProps";
@@ -27,6 +28,9 @@ export function getProperties(
         }
         if (column.showContentAs !== "customContent") {
             hidePropertyIn(defaultProperties, values, "columns", index, "content");
+        }
+        if (column.showContentAs === "customContent") {
+            hidePropertyIn(defaultProperties, values, "columns", index, "tooltip");
         }
         if (!values.columnsSortable) {
             hidePropertyIn(defaultProperties, values, "columns", index, "sortable");
@@ -56,7 +60,6 @@ export function getProperties(
             ]);
         }
     });
-
     if (values.pagination !== "buttons") {
         hidePropertyIn(defaultProperties, values, "pagingPosition");
     }
@@ -125,13 +128,14 @@ export function getProperties(
     return defaultProperties;
 }
 
-export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps => {
+export const getPreview = (values: DatagridPreviewProps, isDarkMode: boolean): StructurePreviewProps => {
     const hasColumns = values.columns && values.columns.length > 0;
     const columnProps: ColumnsPreviewType[] = hasColumns
         ? values.columns
         : [
               {
                   header: "Column",
+                  tooltip: "",
                   attribute: "",
                   width: "autoFit",
                   columnClass: "",
@@ -145,7 +149,8 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                   size: 1,
                   sortable: false,
                   alignment: "left",
-                  sortProperty: "Property"
+                  wrapText: false,
+                  sortProperty: "property"
               }
           ];
     const columns: RowLayoutProps = {
@@ -157,7 +162,12 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                     type: "Container",
                     borders: true,
                     grow: column.width === "manual" && column.size ? column.size : 1,
-                    backgroundColor: values.columnsHidable && column.hidable === "hidden" ? "#F5F5F5" : undefined,
+                    backgroundColor:
+                        values.columnsHidable && column.hidable === "hidden"
+                            ? isDarkMode
+                                ? "#3E3E3E"
+                                : "#F5F5F5"
+                            : undefined,
                     children: [
                         column.showContentAs === "customContent"
                             ? {
@@ -187,7 +197,7 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
     const titleHeader: RowLayoutProps = {
         type: "RowLayout",
         columnSize: "fixed",
-        backgroundColor: "#daeffb",
+        backgroundColor: isDarkMode ? "#3B5C8F" : "#DAEFFB",
         borders: true,
         borderWidth: 1,
         children: [
@@ -198,29 +208,12 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                     {
                         type: "Text",
                         content: "Data grid 2G",
-                        fontColor: "#2074c8"
+                        fontColor: isDarkMode ? "#6DB1FE" : "#2074C8"
                     }
                 ]
             }
         ]
     };
-    // const buttonsHeader: RowLayoutProps = {
-    //     type: "RowLayout",
-    //     columnSize: "fixed",
-    //     borders: true,
-    //     borderWidth: 1,
-    //     children: values.buttons.map((button)=>{
-    //        return {
-    //             type: "Selectable",
-    //             object: button,
-    //             child: {
-    //                 type: 'Container'
-
-    //             }
-
-    //         }
-    //     });
-    // };
     const headerFilters = {
         type: "RowLayout",
         columnSize: "fixed",
@@ -247,7 +240,13 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                             ? column.size
                             : 1
                         : undefined,
-                backgroundColor: isColumnHidden ? "#DCDCDC" : "#F5F5F5",
+                backgroundColor: isColumnHidden
+                    ? isDarkMode
+                        ? "#4F4F4F"
+                        : "#DCDCDC"
+                    : isDarkMode
+                    ? "#3E3E3E"
+                    : "#F5F5F5",
                 children: [
                     {
                         type: "Container",
@@ -258,7 +257,15 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                                 bold: true,
                                 fontSize: 10,
                                 content: column.header ? column.header : "Header",
-                                fontColor: column.header ? undefined : isColumnHidden ? "#DCDCDC" : "#F5F5F5"
+                                fontColor: column.header
+                                    ? undefined
+                                    : isColumnHidden
+                                    ? isDarkMode
+                                        ? "#4F4F4F"
+                                        : "#DCDCDC"
+                                    : isDarkMode
+                                    ? "#3E3E3E"
+                                    : "#F5F5F5"
                             }
                         ]
                     },
@@ -286,6 +293,47 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
                 : content;
         })
     };
+    const button: RowLayoutProps = {
+        type: "RowLayout",
+        columnSize: "grow",
+        backgroundColor: isDarkMode ? "#3B5C8F" : "#DAEFFB",
+        borders: true,
+        borderWidth: 1,
+        children: []
+    };
+    button.children = values.buttons.map(btn => {
+        const container = {
+            type: "Container",
+            borders: true,
+            backgroundColor: isDarkMode ? "#4F4F4F" : "#DCDCDC",
+            children: [] as StructurePreviewProps[]
+        };
+        // if (btn.icon) {
+        //     const iconProp = {
+        //         type: "Image",
+        //     } as ImageProps;
+        //
+        //     switch (btn.icon.type) {
+        //         case "image":
+        //             iconProp.data = btn.icon.imageUrl;
+        //             break;
+        //         case "glyph":
+        //             iconProp.document = btn.icon.iconClass;
+        //             break;
+        //     }
+        //     container.children.push(iconProp);
+        // }
+        container.children.push({
+            type: "Text",
+            bold: true,
+            fontSize: 10,
+            content: `${btn.icon ? "I" : ""} ${btn.caption ? btn.caption : "(No caption)"}`,
+            fontColor: isDarkMode ? "#DCDCDC" : "#4F4F4F"
+        } as TextProps);
+
+        return container;
+    }) as StructurePreviewProps[];
+
     const footer =
         values.showEmptyPlaceholder === "custom"
             ? [
@@ -307,7 +355,7 @@ export const getPreview = (values: DatagridPreviewProps): StructurePreviewProps 
         type: "Container",
         children: [
             titleHeader,
-            // buttonsHeader,
+            button,
             ...(values.showHeaderFilters && values.filterList.length > 0 ? [headerFilters] : []),
             headers,
             ...Array.from({ length: 5 }).map(() => columns),