+++ packages/pluggableWidgets/2G-datagrid-web/src/Datagrid.tsx
@@ -1,10 +1,9 @@
 import { createElement, ReactElement, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { ColumnsType, DatagridContainerProps } from "../typings/DatagridProps";
+import { ButtonsType, ColumnsType, DatagridContainerProps } from "../typings/DatagridProps";
 import { FilterCondition } from "mendix/filters";
 import { and } from "mendix/filters/builders";
 
 import { RemoteSortConfig, Table, TableColumn } from "./components/Table";
-import classNames from "classnames";
 import {
     FilterFunction,
     FilterType,
@@ -12,8 +11,9 @@ import {
     useFilterContext,
     useMultipleFiltering
 } from "@mendix/piw-utils-internal/components/web";
-import { executeAction, isAvailable } from "@mendix/piw-utils-internal";
+import { isAvailable } from "@mendix/piw-utils-internal";
 import { extractFilters } from "./utils/filters";
+import { useCellRenderer } from "./utils/useCellRenderer";
 
 import "./ui/Datagrid.scss";
 
@@ -21,10 +21,6 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
     const id = useRef(`DataGrid${generateUUID()}`);
 
     const [sortParameters, setSortParameters] = useState<{ columnIndex: number; desc: boolean } | undefined>(undefined);
-    const [remoteSortConfig, setRemoteSortConfig] = useState<RemoteSortConfig>({
-        ascending: props.sortAscending?.value,
-        property: props.sortAttribute?.value
-    });
     const isInfiniteLoad = props.pagination === "virtualScrolling";
     const currentPage = isInfiniteLoad
         ? props.datasource.limit / props.pageSize
@@ -33,9 +29,13 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
     const [filtered, setFiltered] = useState(false);
     const multipleFilteringState = useMultipleFiltering();
     const { FilterContext } = useFilterContext();
+    const cellRenderer = useCellRenderer({ columns: props.columns, onClick: props.onClick });
 
     useEffect(() => {
         props.datasource.requestTotalCount(true);
+        if (props.datasource.totalCount) {
+            setTotalCount(props.datasource.totalCount);
+        }
         if (props.datasource.limit === Number.POSITIVE_INFINITY) {
             props.datasource.setLimit(props.pageSize);
         }
@@ -47,46 +47,8 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
         }
     }, [props.datasource, props.configurationAttribute, filtered]);
 
-    useEffect(() => {
-        if (props.sortingType === "remote") {
-            if (
-                props.sortAscending?.value != remoteSortConfig.ascending ||
-                props.sortAttribute?.value != remoteSortConfig.property
-            ) {
-                setRemoteSortConfig({
-                    ascending: props.sortAscending?.value,
-                    property: props.sortAttribute?.value
-                });
-            }
-        }
-    }, [props.sortAscending, props.sortAttribute]);
-    useEffect(() => {
-        if (props.sortingType === "remote") {
-            let changed = false;
-            // check if any property is set
-            if (remoteSortConfig.property) {
-                if (remoteSortConfig.ascending != null && remoteSortConfig.ascending != props.sortAscending?.value) {
-                    props.sortAscending?.setValue(remoteSortConfig.ascending);
-                    changed = true;
-                }
-                if (remoteSortConfig.property && remoteSortConfig.property != props.sortAttribute?.value) {
-                    props.sortAttribute?.setValue(remoteSortConfig.property);
-                    changed = true;
-                }
-            } else {
-                if (props.sortAttribute?.value) {
-                    props.sortAttribute?.setValue(undefined);
-                    changed = true;
-                }
-            }
-            if (changed) {
-                props.onSortChangedAction?.execute();
-            }
-        }
-    }, [remoteSortConfig]);
-
     const setPage = useCallback(
-        computePage => {
+        (computePage: any) => {
             const newPage = computePage(currentPage);
             if (isInfiniteLoad) {
                 props.datasource.setLimit(newPage * props.pageSize);
@@ -96,10 +58,6 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
         },
         [props.datasource, props.pageSize, isInfiniteLoad, currentPage]
     );
-
-    const onConfigurationChange = useCallback(() => {
-        props.onConfigurationChange?.execute();
-    }, [props.onConfigurationChange]);
 
     const customFiltersState = props.columns.map(() => useState<FilterFunction>());
 
@@ -121,7 +79,7 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
         props.datasource.setFilter(viewStateFilters.current);
     }
 
-    if (sortParameters && props.sortingType === "local") {
+    if (props.sortingType === "local" && sortParameters) {
         props.datasource.setSortOrder([
             [props.columns[sortParameters.columnIndex].attribute!.id, sortParameters.desc ? "desc" : "asc"]
         ]);
@@ -149,34 +107,87 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
             ),
         [props.filterList, viewStateFilters.current]
     );
+    /** Remote sorting
+     *
+     */
+    const [remoteSortConfig, setRemoteSortConfig] = useState<RemoteSortConfig>({
+        ascending: props.sortAscending?.value,
+        property: props.sortAttribute?.value
+    });
+    useEffect(() => {
+        if (props.sortingType === "remote") {
+            if (
+                props.sortAscending?.value !== remoteSortConfig.ascending ||
+                props.sortAttribute?.value !== remoteSortConfig.property
+            ) {
+                setRemoteSortConfig({
+                    ascending: props.sortAscending?.value,
+                    property: props.sortAttribute?.value
+                });
+            }
+        }
+    }, [props.sortAscending, props.sortAttribute]);
+    useEffect(() => {
+        if (props.sortingType === "remote") {
+            let changed = false;
+            // check if any property is set
+            if (remoteSortConfig.property) {
+                if (remoteSortConfig.ascending != null && remoteSortConfig.ascending !== props.sortAscending?.value) {
+                    props.sortAscending?.setValue(remoteSortConfig.ascending);
+                    changed = true;
+                }
+                if (remoteSortConfig.property && remoteSortConfig.property !== props.sortAttribute?.value) {
+                    props.sortAttribute?.setValue(remoteSortConfig.property);
+                    changed = true;
+                }
+            } else {
+                if (props.sortAttribute?.value) {
+                    props.sortAttribute?.setValue(undefined);
+                    changed = true;
+                }
+            }
+            if (changed) {
+                props.onSortChangedAction?.execute();
+            }
+        }
+    }, [remoteSortConfig]);
+    /**
+     * End Remote Sorting
+     */
+    const [totalCount, setTotalCount] = useState<number>();
 
+    // /**
+    //  * 2G custom button state
+    //  */
+    // const [selection, setSelection] = useState<ObjectItem[]>([]);
+    // /**
+    //  * Update the selected ObjectItme list.
+    //  * @param item
+    //  */
+    // const updateSelection = (item: ObjectItem): void => {
+    //     switch (props.selectionMode) {
+    //         case "single":
+    //             setSelection([item]);
+    //             break;
+    //         case "multi":
+    //             const selectedItem = selection.indexOf(item);
+    //             let newSelection: ObjectItem[] = [];
+    //             if (selectedItem > -1) {
+    //                 newSelection = selection.filter(x => x.id !== item.id);
+    //             } else {
+    //                 newSelection = [...selection, item];
+    //             }
+    //             for (const select of newSelection) {
+    //                 console.log(select.id);
+    //             }
+    //             setSelection(newSelection);
+    //             break;
+    //     }
+    // };
     return (
         <Table
-            cellRenderer={useCallback(
-                (renderWrapper, value, columnIndex) => {
-                    const column = props.columns[columnIndex];
-                    let content;
-
-                    if (column.showContentAs === "attribute") {
-                        content = <span className="td-text">{column.attribute?.get(value)?.displayValue ?? ""}</span>;
-                    } else if (column.showContentAs === "dynamicText") {
-                        content = <span className="td-text">{column.dynamicText?.get(value)?.value ?? ""}</span>;
-                    } else {
-                        content = column.content?.get(value);
-                    }
-
-                    return renderWrapper(
-                        content,
-                        classNames(`align-column-${column.alignment}`, column.columnClass?.get(value)?.value),
-                        props.onTrigger ? () => executeAction(props.onTrigger?.get(value)) : undefined,
-                        props.defaultTrigger
-                    );
-                },
-                [props.columns, props.defaultTrigger, props.onTrigger]
-            )}
+            cellRenderer={cellRenderer}
             className={props.class}
-            buttons={props.buttons}
-            tableLabel={props.tableLabel}
             columns={columns}
             columnsDraggable={props.columnsDraggable}
             columnsFilterable={props.columnsFilterable}
@@ -185,7 +196,7 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
             columnsSortable={props.columnsSortable}
             data={props.datasource.items ?? []}
             emptyPlaceholderRenderer={useCallback(
-                renderWrapper =>
+                (renderWrapper: any) =>
                     props.showEmptyPlaceholder === "custom" ? renderWrapper(props.emptyPlaceholder) : <div />,
                 [props.emptyPlaceholder, props.showEmptyPlaceholder]
             )}
@@ -243,14 +254,12 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
                 [FilterContext, customFiltersState, filterList, multipleInitialFilters, props.filtersPlaceholder]
             )}
             id={id.current}
-            selectionMode={props.selectionMode}
-            numberOfItems={props.datasource.totalCount}
-            onSettingsChange={props.onConfigurationChange ? onConfigurationChange : undefined}
+            numberOfItems={totalCount}
             page={currentPage}
             pageSize={props.pageSize}
             paging={props.pagination === "buttons"}
             pagingPosition={props.pagingPosition}
-            rowClass={useCallback(value => props.rowClass?.get(value)?.value ?? "", [props.rowClass])}
+            rowClass={useCallback((value: any) => props.rowClass?.get(value)?.value ?? "", [props.rowClass])}
             setPage={setPage}
             setSortParameters={setSortParameters}
             settings={props.configurationAttribute}
@@ -262,10 +271,34 @@ export default function Datagrid(props: DatagridContainerProps): ReactElement {
                 },
                 [props.columns]
             )}
+            /**
+             * Custom 2G props
+             */
+            buttons={transformButtonsType(props.buttons)}
+            selectionMode={props.selectionMode}
+            tableLabel={props.tableLabel}
             remoteSortConfig={remoteSortConfig}
             setRemoteSortConfig={setRemoteSortConfig}
+            defaultTrigger={props.defaultTrigger}
         />
     );
+}
+
+export interface ButtonsTypeExt extends ButtonsType {
+    key: number;
+}
+function transformButtonsType(buttons: ButtonsType[]): ButtonsTypeExt[] {
+    return buttons.map(btn => {
+        return {
+            key: buttons.indexOf(btn),
+            action: btn.action,
+            buttonStyle: btn.buttonStyle,
+            caption: btn.caption,
+            actionNoContext: btn.actionNoContext,
+            icon: btn.icon,
+            renderMode: btn.renderMode
+        } as ButtonsTypeExt;
+    });
 }
 
 function transformColumnProps(props: ColumnsType[]): TableColumn[] {